[binding] Start old binding rewrite.
